requirements.txt
certifi
charset-normalizer
contourpy
cycler
fonttools
idna
kiwisolver
matplotlib
numpy
packaging
Pillow==9.5.0
pyparsing
python-dateutil
python-dotenv
requests
six
urllib3
wordcloud
pytrends
google-api-python-client
streamlit==1.24.1
SQLAlchemy
plotly
duckduckgo_search
pandas>=1.5.3


